requirements.txt
# Previously
# Flask==2.0.2
# streamlit==1.8.1
# Pillow==8.4.0

# New
torch==1.12.0
torchvision==0.13.0
Pillow==9.3.0
opencv-python == 4.6.0.66
numpy==1.23.4
python-multipart==0.0.5
